% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clic_lookup.R
\name{clic_lookup}
\alias{clic_lookup}
\title{Fetch listing of CLiC content}
\usage{
clic_lookup()
}
\value{
Returns a \code{data.frame} listing the texts for each of the available corpora.
}
\description{
Fetches a listing of the texts for each of the available corpora in CLiC.
}
