% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clic_subset.R
\name{clic_subset}
\alias{clic_subset}
\title{Fetch subsets.}
\usage{
clic_subset(corpora, subset, contextsize = 3, metadata = FALSE, json = FALSE)
}
\arguments{
\item{corpora}{1+ corpus name (e.g. 'dickens') or book name ('AgnesG') to search within}

\item{subset}{subset to search through, one of shortsus/longsus/nonquote/quote/all. Default 'all' (i.e. all text)}

\item{contextsize}{size of context window around search results. Default 3.}

\item{metadata}{Return metadata. TRUE/FALSE}

\item{json}{JSON format. TRUE/FALSE}
}
\value{
A \code{data.frame} with one entry per result.
}
\description{
Subsets of given texts, for example quotations.
}
