\name{clicclient-package}
\alias{clicclient-package}
\alias{clicclient}
\docType{package}

\title{Client for the CLiC API.}
\description{
    A set of function wrappers for the CLiC API.
}
\details{
    The CLiC web interface can be found at \url{http://clic.bham.ac.uk}.
    CLiC also provides an \sQuote{application programming interface} (API)
    that makes the data used by the web interface available in a machine
    readable form.
    This package offers a set of wrapper functions to facilitate
    interaction with the API from R.

    For a list of all documentation use \code{library(help="clicclient")}.
}

\author{
    Maintainer: Anthony Hennessey <anthony.hennessey@nottingham.ac.uk>.
}
