% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clicclient.R
\docType{package}
\name{clicclient}
\alias{clicclient}
\title{clicclient: Client for the CLiC API.}
\description{
A set of function wrappers for the CLiC API.
}
\details{
The CLiC web interface can be found at \url{http://clic.bham.ac.uk}.
CLiC also provides an \sQuote{application programming interface} (API)
that makes the data used by the web interface available in a machine
readable form.

This package offers a set of wrapper functions to facilitate
interaction with the API from R.

For a list of all documentation use \code{library(help="clicclient")}.
}
