% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clic_subset.R
\name{subset_df}
\alias{subset_df}
\title{One row of the subsets data frame.}
\usage{
subset_df(response_data, metadata = FALSE)
}
\arguments{
\item{response_data}{list of subset data from clic subset endpoint}

\item{metadata}{TRUE/FALSE}
}
\value{
a data frame with character variables left, node, right and book
if metadata is TRUE then the output also has
numeric variables chapter, paragraph, sentence, begin and end.
}
\description{
Builds a data frame corresponding to a single
response from the subset endpoint.
}
\keyword{internal}
