% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clic_subset.R
\name{subset_query_list}
\alias{subset_query_list}
\title{A query list for subset queries.}
\usage{
subset_query_list(params)
}
\arguments{
\item{params}{list of parameters}
}
\value{
query list for subset endpoint
}
\description{
Builds a query list to pass to clic_request when
calling the subset endpoint.
}
\keyword{internal}
