\name{clic_lookup}
\alias{clic_lookup}

\title{Fetch listing of CLiC content}
\description{
    Fetches a listing of the texts for each of the available corpora in CLiC.

}

\usage{
    clic_lookup() 
}

\value{
    Returns a \code{data.frame} listing the texts for each of the available corpora.
}

