% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clic_set_user_agent.R
\name{clic_set_user_agent}
\alias{clic_set_user_agent}
\alias{clic_get_user_agent}
\title{Get/Set User-Agent}
\usage{
clic_set_user_agent(x)

clic_get_user_agent()
}
\arguments{
\item{x}{The new \sQuote{User-Agent} string.}
}
\value{
Both functions return the User-Agent string.
}
\description{
Get or set the \sQuote{User-Agent} string for CLiC API requests.
}
\details{
Setting the User-Agent helps us to understand who is using the API.
Ideally set the User-Agent to something that either identifies you
or your application.

The string you supply will be concatenated with the version details of
this package. The concatenated string is returned.
}
